requirements.txt
anyio==3.6.1
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
filelock==3.7.0
google-auth==2.6.6
h11==0.13.0
idna==3.3
kubernetes==23.6.0
oauthlib==3.2.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
shortuuid==1.0.9
six==1.16.0
sniffio==1.2.0
starlette==0.20.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
websocket-client==1.3.2
